refactor: move task response and task processor to individual file
